backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]>=3.0.0

# Authentication
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0

# ML dependencies (for future weeks)
numpy==1.26.4
opencv-python==4.10.0.84

# Testing
requests==2.32.3
